ui tweaks
